// src/enums/productId.ts
export enum PRODUCT_ID {
  EXPENSE = "expense",
  FITPASS = "fitpass",
  GO_FOR_EAT_AND_GIFT = "go-for-eat-and-gift",
  CHEQUE_DEJEUNER = "cheque-dejeuner",
  MERCHANT_MEAL_ACCEPTANCE = "merchant-meal-acceptance",
  MERCHANT_NON_MEAL_ACCEPTANCE = "merchant-non-meal-acceptance",
  MERCHANT_FITPASS_ACCEPTANCE = "merchant-fitpass-acceptance",
  FLEXONE_SUBSCRIPTION = "flexone-subscription",
  FLEXONE_COMMISSION = "flexone-commission",
  REWARDS = "rewards",
  GIFT = "gift",
}